fixed edit sound file

+++ modules/sound/traits/SoundFile.php
@@ -5,22 +5,21 @@ trait SoundFile {
 	public function saveFile()
 	{
 		if(empty($_FILES) || $_FILES['file']['error'] != 0) return false;
-		$id_last = self::getLastId('sounds');
-		$file = $this->setFileName($id_last);
+		$file = $this->setFileName();
 		if ($this->moveFile($file)) return $file;
 	}
 	
 	private function moveFile($file)
 	{
 		$destination = './web/sounds/'.$file;
+		if (file_exists($destination)) return $this->saveFile();
 		return move_uploaded_file($_FILES['file']['tmp_name'], $destination);
 	}
 	
-	private function setFileName($id_last)
+	private function setFileName()
 	{
-		$id_last = $id_last ? $id_last : 0;
 		$extension_file = explode('.', $_FILES['file']['name'])[1];
-		$filename = $id_last + 1;
+		$filename = time();
 		return $filename.'.'.$extension_file;
 	}
